move auth handlers to separate pkg
