+++ tools/http-test-helper/src/main/java/org/enso/shttp/HTTPTestHelperServer.java
@@ -10,6 +10,9 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
 import java.util.stream.Stream;
+
+import org.enso.shttp.auth.BasicAuthTestHandler;
+import org.enso.shttp.auth.TokenAuthTestHandler;
 import org.enso.shttp.cloud_mock.CloudRoot;
 import sun.misc.Signal;
 import sun.misc.SignalHandler;
