+++ tools/http-test-helper/src/main/java/org/enso/shttp/auth/BasicAuthTestHandler.java
@@ -1,6 +1,8 @@
-package org.enso.shttp;
+package org.enso.shttp.auth;
 
 import com.sun.net.httpserver.HttpExchange;
+import org.enso.shttp.SimpleHttpHandler;
+
 import java.io.IOException;
 import java.util.Base64;
 import java.util.List;
