+++ tools/http-test-helper/src/main/java/org/enso/shttp/auth/TokenAuthTestHandler.java
@@ -1,4 +1,4 @@
-package org.enso.shttp;
+package org.enso.shttp.auth;
 
 import com.sun.net.httpserver.HttpExchange;
 import java.io.IOException;
